rip out void on journal
