+++ src/Book.ts
@@ -9,6 +9,9 @@ import { JournalNotFoundError } from "./errors/JournalNotFoundError";
 import { BookConstructorError } from "./errors/BookConstructorError";
 import { lockModel } from "./models/lock";
 import { getBestSnapshot, IBalance, snapshotBalance } from "./models/balance";
+import { JournalAlreadyVoidedError, MediciError } from ".";
+import { handleVoidMemo } from "./helper/handleVoidMemo";
+import { addReversedTransactions } from "./helper/addReversedTransactions";
 
 export class Book<U extends ITransaction = ITransaction, J extends IJournal = IJournal> {
   name: string;
@@ -179,7 +182,46 @@ export class Book<U extends ITransaction = ITransaction, J extends IJournal = IJ
       throw new JournalNotFoundError();
     }
 
-    return journal.void(this, reason, options);
+    if (journal.voided) {
+      throw new JournalAlreadyVoidedError();
+    }
+  
+    reason = handleVoidMemo(reason, journal.memo);
+  
+    // Not using options.session here as this read operation is not necessary to be in the ACID session.
+    const transactions = await transactionModel.collection.find({ _journal: journal._id }).toArray();
+    if (transactions.length === 0)
+      throw new MediciError(`Transactions for journal ${journal._id} not found on book ${journal.book}`);
+  
+    const entry = this.entry(reason, null, journal._id);
+  
+    addReversedTransactions(entry, transactions as ITransaction[]);
+  
+    // Set this journal to void with reason and also set all associated transactions
+    const resultOne = await journalModel.collection.updateOne(
+      { _id: journal._id },
+      { $set: { voided: true, void_reason: reason } },
+      {
+        session: options.session, // We must provide either session or writeConcern, but not both.
+        writeConcern: options.session ? undefined : { w: 1, j: true }, // Ensure at least ONE node wrote to JOURNAL (disk)
+      }
+    );
+    if (!resultOne.acknowledged) throw new MediciError(`Failed to void ${journal.memo} journal on book ${journal.book}`);
+    if (resultOne.modifiedCount === 0) throw new MediciError(`Already voided ${journal.memo} journal on book ${journal.book}`);
+  
+    const resultMany = await transactionModel.collection.updateMany(
+      { _journal: journal._id },
+      { $set: { voided: true, void_reason: reason } },
+      {
+        session: options.session, // We must provide either session or writeConcern, but not both.
+        writeConcern: options.session ? undefined : { w: 1, j: true }, // Ensure at least ONE node wrote to JOURNAL (disk)
+      }
+    );
+    if (!resultMany.acknowledged) throw new MediciError(`Failed to void ${journal.memo} transactions on book ${journal.book}`);
+    if (resultMany.modifiedCount !== transactions.length)
+      throw new MediciError(`Already voided ${journal.memo} transactions on book ${journal.book}`);
+  
+    return entry.commit(options);
   }
 
   async writelockAccounts(accounts: string[], options: Required<Pick<IOptions, "session">>): Promise<Book<U, J>> {
